Support PHPStan 1.*
