+++ src/Carbon/PHPStan/Macro.php
@@ -13,147 +13,44 @@ declare(strict_types=1);
 
 namespace Carbon\PHPStan;
 
-use Closure;
 use PHPStan\Reflection\Php\BuiltinMethodReflection;
-use PHPStan\TrinaryLogic;
-use ReflectionClass;
-use ReflectionFunction;
 use ReflectionMethod;
-use ReflectionParameter;
-use ReflectionType;
-use stdClass;
-use Throwable;
 
-final class Macro implements BuiltinMethodReflection
-{
-    /**
-     * The class name.
-     *
-     * @var class-string
-     */
-    private $className;
+$method = new ReflectionMethod(BuiltinMethodReflection::class, 'getFileName');
 
-    /**
-     * The method name.
-     *
-     * @var string
-     */
-    private $methodName;
-
-    /**
-     * The reflection function/method.
-     *
-     * @var ReflectionFunction|ReflectionMethod
-     */
-    private $reflectionFunction;
-
-    /**
-     * The parameters.
-     *
-     * @var ReflectionParameter[]
-     */
-    private $parameters;
-
-    /**
-     * The is static.
-     *
-     * @var bool
-     */
-    private $static = false;
-
-    /**
-     * Macro constructor.
-     *
-     * @param string $className
-     * @phpstan-param class-string $className
-     *
-     * @param string   $methodName
-     * @param callable $macro
-     */
-    public function __construct(string $className, string $methodName, $macro)
+if ($method->hasReturnType()) {
+    final class Macro extends AbstractMacro
     {
-        $this->className = $className;
-        $this->methodName = $methodName;
-        $this->reflectionFunction = \is_array($macro)
-            ? new ReflectionMethod($macro[0], $macro[1])
-            : new ReflectionFunction($macro);
-        $this->parameters = $this->reflectionFunction->getParameters();
+        /**
+         * {@inheritdoc}
+         */
+        public function getFileName(): string
+        {
+            return $this->reflectionFunction->getFileName();
+        }
 
-        if ($this->reflectionFunction->isClosure()) {
-            try {
-                $closure = $this->reflectionFunction->getClosure();
-                $boundClosure = Closure::bind($closure, new stdClass());
-                $this->static = (!$boundClosure || (new ReflectionFunction($boundClosure))->getClosureThis() === null);
-            } catch (Throwable $e) {
-                $this->static = true;
-            }
+        /**
+         * {@inheritdoc}
+         */
+        public function getStartLine(): ?int
+        {
+            return $this->reflectionFunction->getStartLine();
+        }
+
+        /**
+         * {@inheritdoc}
+         */
+        public function getEndLine(): ?int
+        {
+            return $this->reflectionFunction->getEndLine();
         }
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function getDeclaringClass(): ReflectionClass
-    {
-        return new ReflectionClass($this->className);
-    }
+    return;
+}
 
-    /**
-     * {@inheritdoc}
-     */
-    public function isPrivate(): bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isPublic(): bool
-    {
-        return true;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isFinal(): bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isInternal(): bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isAbstract(): bool
-    {
-        return false;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isStatic(): bool
-    {
-        return $this->static;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getDocComment(): ?string
-    {
-        return $this->reflectionFunction->getDocComment() ?: null;
-    }
-
+final class Macro extends AbstractMacro
+{
     /**
      * {@inheritdoc}
      *
@@ -162,30 +59,6 @@ final class Macro implements BuiltinMethodReflection
     public function getFileName()
     {
         return $this->reflectionFunction->getFileName();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getName(): string
-    {
-        return $this->methodName;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getParameters(): array
-    {
-        return $this->parameters;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getReturnType(): ?ReflectionType
-    {
-        return $this->reflectionFunction->getReturnType();
     }
 
     /**
@@ -206,42 +79,5 @@ final class Macro implements BuiltinMethodReflection
     public function getEndLine()
     {
         return $this->reflectionFunction->getEndLine();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isDeprecated(): TrinaryLogic
-    {
-        return TrinaryLogic::createFromBoolean(
-            $this->reflectionFunction->isDeprecated() ||
-            preg_match('/@deprecated/i', $this->getDocComment() ?: '')
-        );
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isVariadic(): bool
-    {
-        return $this->reflectionFunction->isVariadic();
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getPrototype(): BuiltinMethodReflection
-    {
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getReflection(): ?ReflectionMethod
-    {
-        return $this->reflectionFunction instanceof ReflectionMethod
-            ? $this->reflectionFunction
-            : null;
     }
 }
